requirements.txt
PyQt6>=6.6.0
pyserial>=3.5
python-dotenv>=1.0.0
requests>=2.31.0
